requirements.txt
requests
python-dotenv
azure-functions
azure-storage-blob
